requirements.txt
alembic==1.4.2
astroid==2.4.2
azure-core==1.7.0
azure-storage-blob==12.3.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
databricks-cli==0.11.0
docker==4.2.2
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.7
gorilla==0.3.0
idna==2.10
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mlflow==1.10.0
msrest==0.6.17
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
pandas==1.0.3
Pillow==7.2.0
prometheus-client==0.8.0
prometheus-flask-exporter==0.15.4
protobuf==3.12.2
pycparser==2.20
pylint==2.5.3
pyparsing==2.4.6
pypiwin32==223
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
pywin32==228
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
sip==4.19.13
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
toml==0.10.1
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.10
waitress==1.4.4
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
